Get connections from the pool in the correct way
Passing an acquire() object I suspect was destroying all the connections in the pool, and causing the bot to run out of said connections and crash. Putting them in "async with" blocks is definitely better management of the pool, but I am unsure of whether this is the actual solution.