Add Automatic PvE Feature
Adds the option to make PvE automatic. In this case, player attacks are chosen randomly and play is updated every 3 seconds. Gold and xp rewards are reduced by 75%. Playing interactively remains the default option.
This is implemented by adding the str auto which is "Y" if player wishes to play interactively, and "N" otherwise.